delete authSlice and modified loginrequest,loginPage
