+++ src/app/login/page.tsx
@@ -1,22 +1,15 @@
 "use client"
-import { useDispatch, useSelector } from "react-redux"
-import { RootState, AppDispatch } from "@/redux/store"
+import { useState, useTransition } from "react"
 import { useRouter } from "next/navigation"
 import { useForm } from "react-hook-form"
 import { loginRequest } from "@/services/loginRequest"
-import { handleLogin } from "@/redux/states/user"
-import { LoginResponse } from "@/redux/states/user"
-
-export interface LoginUser {
-  user: string
-  password: string
-}
+import { LoginUser } from "@/services/loginRequest"
 
 const LoginPage: React.FC = () => {
-  const router = useRouter()
+  const [error, setError] = useState("")
+  const [isPending, startTransition] = useTransition()
 
-  const dispatch = useDispatch<AppDispatch>()
-  const { loading, error } = useSelector((state: RootState) => state.auth)
+  const router = useRouter()
 
   const {
     register,
@@ -24,26 +17,23 @@ const LoginPage: React.FC = () => {
     formState: { errors },
   } = useForm<LoginUser>()
 
-  const onSubmit = handleSubmit(async (data: LoginUser) => {
-    try {
-      const response: LoginResponse = await dispatch(
-        loginRequest(data)
-      ).unwrap()
-
-      dispatch(
-        handleLogin({ success: response.success, userType: response.userType })
-      )
-      router.push("/precios")
-    } catch (error) {
-      console.log(error)
-    }
-  })
+  const onSubmit = (data: LoginUser) => {
+    startTransition(async () => {
+      try {
+        await loginRequest(data)
+        router.push("/precios")
+      } catch (error: any) {
+        console.log(error)
+        setError(error.toString())
+      }
+    })
+  }
 
   return (
     <div className="min-h-screen flex items-center justify-center">
-      <form onSubmit={onSubmit} className="flex flex-col gap-2">
-        {error && <p>{JSON.stringify(error)}</p>}
+      <form onSubmit={handleSubmit(onSubmit)} className="flex flex-col gap-2">
         {errors.user && <p className="text-red-500">{errors.user.message}</p>}
+
         <div>
           <input
             type="text"
@@ -53,7 +43,7 @@ const LoginPage: React.FC = () => {
               required: { value: true, message: "*Campos obligatorios" },
               maxLength: {
                 value: 10,
-                message: "*Maximo de 10 caracteres",
+                message: "*Máximo de 10 caracteres",
               },
             })}
           />
@@ -69,7 +59,7 @@ const LoginPage: React.FC = () => {
           />
         </div>
         <button className="bg-indigo-500 p-2 rounded-md" type="submit">
-          {loading ? "Entrando.." : "Entrar"}
+          {!isPending ? "Entrar" : "Entrando.."}
         </button>
       </form>
     </div>
