+++ src/config/cookies.ts
@@ -1,7 +1,7 @@
 "use server"
 import { cookies } from "next/headers"
 
-export const setToken = async (token: any) => {
+export const setToken = async (token: string) => {
   const nextCookie = cookies()
   return nextCookie.set("userToken", token)
 }
