+++ src/redux/store.ts
@@ -1,10 +1,8 @@
 import { configureStore } from "@reduxjs/toolkit"
 import priceReducer from "./states/price"
-import authReducer from "./states/user"
 
 export const store = configureStore({
   reducer: {
-    auth: authReducer,
     price: priceReducer,
   },
   middleware: (getDefaultMiddleware) => getDefaultMiddleware(),
