+++ src/services/loginRequest.ts
@@ -1,31 +1,34 @@
-import { createAsyncThunk } from "@reduxjs/toolkit"
 import { AppError } from "./errorRequest"
-import { LoginUser } from "@/app/login/page"
-import { LoginResponse } from "@/redux/states/user"
 import { setToken } from "@/config/cookies"
 import { setUserType } from "@/config/setUserType"
 
-export const loginRequest = createAsyncThunk<LoginResponse, LoginUser>(
-  "user/loginRequest",
-  async (user, { rejectWithValue }) => {
-    const baseUrl = process.env.NEXT_PUBLIC_LOCALHOST
+export interface LoginUser {
+  user: string
+  password: string
+}
 
-    await new Promise((resolve) => setTimeout(resolve, 3000))
+export interface LoginResponse {
+  success: string
+  userType: string
+}
+export const loginRequest = async (user: LoginUser): Promise<LoginResponse> => {
+  const baseUrl = process.env.NEXT_PUBLIC_LOCALHOST
 
-    const response = await fetch(`${baseUrl}/login`, {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      credentials: "include",
-      body: JSON.stringify(user),
-    })
-    const data = await response.json()
+  await new Promise((resolve) => setTimeout(resolve, 3000))
 
-    await JSON.stringify([setToken(data.success), setUserType(data.userType)])
+  const response = await fetch(`${baseUrl}/login`, {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    credentials: "include",
+    body: JSON.stringify(user),
+  })
 
-    if (!response.ok) {
-      const error = new AppError(data.error, data.message, data.status)
-      return rejectWithValue(error)
-    }
-    return data
+  const data = await response.json()
+
+  await JSON.stringify([setToken(data.success), setUserType(data.userType)])
+
+  if (!response.ok) {
+    throw new AppError(data.error, data.message, data.status)
   }
-)
+  return data
+}
